Fix Ordering By Time Bug
Previously we moved files to a tmp folder when renaming to avoid
renameing to an already existing file name so we don't overwrite files
accidentally, then we reread the files from that tmp folder and ordered
the files as we read them by time. This assumes that the files are read
in alphabetic order by this is not true. In this fix the files' atime
and mtime are changed when we rename them alphabetically so the
timestamps ordering matchs the alphabetic ordering.